New method: flexMap
An improved version of the map method, which outputs an array
with less or more items than the original one. It will be
passed four arguments to the supplied callback:

1) the current array item
2) the index
3) the array instance
4) an object holding two operation methods (skip and many).

You can use the ops.skip operation to ignore a given item (so it
won't have a matching item in the output array). You can also
map from one to multiple items by using the ops.many operation.